Upload V0.1
Solution: The goal of this project is to classify the complaint into a specific product category. Since it has multiple categories, it becomes a multiclass classification that can be solved through many machine learning algorithms.
Once the algorithm is in place, whenever there is a new complaint, we can easily categorize it and can then be redirected it to the concerned person. This will save a lot of time because we are minimizing the human interventions to decide whom this complaint should go to.